requirements.txt
flask
openai
jwt
pandas
qdrant-client
langchain-openai
cohere
python-dotenv
